requirements.txt
Flask>=3.1.3
python-dotenv>=1.0.1
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.7
PyMySQL>=1.1.1
Flask-Limiter>=3.5.0
